refactor: make interaction availability as extension function
